Avoid writing the 1984-era 7702 rate literally
Now that the 7702 interest rate is no longer a constant, the original
1984 rate must not be written literally anywhere. Examined every
occurrence of /\.0[46][^0-9]/ and /[46]%/ to make sure nothing was
missed in commit 7930f69291 of 20201229T0001Z. Abstracted such interest
rates in the few cases found (one unit test, and several comments) even
though they're of no material import. Changed local variables' names in
another unit test because the GLP and GSP rates are not necessarily
"4" or "6" percent since 2021 because of H.R. 6800.