Resolve defect markers imported with 7702 specifications
It is better to mark defects in code, where they're likelier to be
noticed and fixed. Some of the "defects" marked in this document,
which was copied verbatim from the lmi webpages repository, where it
had hardly been touched for two decades, don't seem so defective now.
Disposition of each:

  <!-- TODO ?? Move: this should refer to all assumptions ...

That's a valid drafting suggestion, but it pertains to the document only
and does not indicate any lmi shortcoming. Therefore, changed the marker
to one less severe.

  <!-- TODO ?? Don&rsquo;t permit SA to decrease below the current
       corridor DB? -->

This is a note-to-self from the turn of the century. Presumably it would
have been dismissed then, had time been taken to talk it through. There
is no evident need or justification for such a rule, and no realistic
hope of imposing it anyway.

  <!-- TODO ?? But at the moment <tt>lmi</tt> seems to track the
       lowest benefit since ...

Moved to 'ihs_irc7702.cpp', where it's more likely to be addressed; and
demoted in severity, because its new context suggests that the issue may
have been resolved in favor of doing what the comment describes.

  <!-- TODO ?? merge later reference? -->

This document underwent many cycles of review and revision, of which
this comment seems to be artifact: page 1438 of the OBRA House Report is
cited in the immediately preceding footnote (with which it would not
seem appropriate to merge this one), but nowhere else.

  MEC as of the (past) failure date, but [TODO ?? cite
  anticipation rules described elsewhere].

Resolved by citing 7702A(d).