Perform a transcendental calculation OAOO: part deux
Removed old code that had been kept only to validate its replacement.

Speed is nearly unaffected, which is somewhat surprising: calling
i_upper_12_over_12_from_i<double>() for all durations in class
InterestRates, with the important optimization of reusing the previous
year's rate if it's the same instead of recalculating it, should be
faster than calling it for each year in the monthiversary loop with no
such optimization. OTOH, some of the new code such as (simplified):

    for(int j = mce_gen_curr; j < mc_n_gen_bases; ++j)
        for(int k = mce_sep_full; k < mc_n_sep_bases; ++k)
            SepAcctGrossRate_[i]   [k] = Zero_;

is certainly inefficient in that identical assignments are performed
mc_n_gen_bases times when once would suffice.

Nevertheless, this is a good change because it moves a compound-interest
calculation from class AccountValue to class InterestRates where such
calculations naturally belong, and it proceeds naturally from commit
365cbb477dfdc (whose more palpable speedup is recorded in 1017bcb08614).