Assume genacct and loan rates are always needed [312]
US states generally require policy to include a loan provision, so
loans are always needed. Loaned funds are usually held in the general
account, so a general account is always needed, even if only for loans.

Other types of interest rates are needed only conditionally, and some
cost is saved by not calculating them when they're not needed; but these
two categories are always needed.