Prepare to resolve some marked defects
Markers attribute discrepancies in the lengths of certain vectors to the
antediluvian branch, but that branch has long since used the same input
class as production, so that diagnosis has become incorrect. The vectors
in question are of unexpected length (zero) only when initialized by the
BasicValues ctor used by the 7702 server, so the root cause is resolved
by amending that ctor.

For the moment, it is merely asserted that simplification will work.
The actual simplification will follow in the next commit.