Clarify lmi's 7702 treatment of flat extras
The original wording:
  for calculating UL &sect;7702 and &sect;7702A limits
was more restrictive than the replacement:
  for all &sect;7702 and &sect;7702A purposes
leading the reader to wonder what was intended to be excluded.
Considering all possibilities:
  {GLP, GSP, 7PP, NSPs, CVAT corridor, DCV}
the first three are clearly "limits", and the next two are specified
to use the same 7702 mortality; but so is the last:
  Update DCV for CVAT contracts. Accumulate it like
  AV, but using §7702 assumptions.
so this clarification does not actually change the meaning.

Furthermore, this change is consonant with [8.4]:
  We may indeed choose to use 100% of safe-harbor mortality in all cases
  as an administrative simplification, in which case changes in table
  ratings or flat extras are not adjustment events; that is the only
  option lmi currently supports.