Clarify DB_AnnInterestRate7702
H.R. 6800 calls this the "applicable accumulation test minimum rate",
which is unfortunate because it's the everything-but-GSP rate.

The GSP rate is "the applicable accumulation test minimum rate (as
defined in subsection (b)(3)) plus 2 percentage points".

Thus, there are two interest rates, X and X+0.02, and it suffices to
store only X in the database.