Resurrect ancient code for 7702 (loan) interest
View this revision with '--ignore-all-space' to see clearly that it
changes the condition on a block, from unconditional suppression with
'/*...*/', to conditional with 'if(condition) {}', depending on the
recently added database entity DB_IgnoreLoanRateFor7702.

Although the resurrected code (which has been suppressed since before
the lmi epoch) contains a (marked) defect, this commit has no effect on
regression tests because it suppresses DB_IgnoreLoanRateFor7702 for the
nonce.