Remove a duplicative defect marker [300]
The comment preceding
  stratified_charges::minimum_tiered_spread_for_7702()
already names other items that should be included. One of those items
is "DB_CurrAcctValLoad"; the comment removed hereby mentioned
"comp (sometimes)", but to the extent that's relevant it would be
included in DB_CurrAcctValLoad: compensation per se is not includible,
but the load that provides for it is.