Turn a local variable into a member
The new member is a unique_ptr for concinnity with its congeners. The
definition of pointed-to class i7702 therefore must appear wherever its
dtor is needed, but is not required in the header of the class that owns
the pointer.

Incidentally made the new member a unique_ptr to const, and similarly
constified congeners on which non-const member functions aren't called.