Support a further "refactoring"
This resembles other recent refactorings, but its assertions do not
prove that the code to be retired is equivalent to its replacement:
it cannot be, because one ignores loan rates unconditionally while the
other does not. It passes regression tests only because such tests are
very limited for '.mec' input files. The '.gpt' testcases are better,
but they bypass this code because test_one_days_gpt_transactions()
returns early.

Any inequivalence is however accidental: the code that will soon be
removed here (#if 1...endif // 1) was intended to be a faithful copy
of the more authoritative code that has become class i7702, so the
forthcoming replacement will actually be an improvement rather than
a refactoring.