Almost add CurrAcctValLoad to 7702 "spread"
This commit uses new code, zeroing out the effect of CurrAcctValLoad
to make acceptance testing easy. The next commit will un-zero it back
in, causing regressions that should be straightforward to test.

The new code explicitly imposes the separate-account-only condition
that is now stated in the 7702 documentation as well, which was
mentioned in a comment added in the preceding commit.