Deduct all asset-based loads as appropriate for 7702 [293]
The expunged comment about "CurrSepAcctLoadBandedByPrem" was wrong:
that load is indeed based on assets. Numerically, it is 'k * assets',
where 'k' is a scalar constant determined by premium, but here premium
is disregarded and the lowest value of 'k' is used. A load that is
calculated as a constant times assets is an asset-based load.