Delete fs::ifstream and ofstream move ctors
These ctors don't exist because they're deleted in the corresponding
base classes, and explicitly defaulting them is misleading (because it
might lead one to believe that they are available, when this is not the
case) and results in -Wdefaulted-function-deleted from clang.

Just delete them explicitly for clarity and to avoid the warnings.