Remove pragma GCC diagnostic push without matching pop
This seems like a mistake and while it is currently harmless to have
such an orphan push, it seems better not to have it at all.

Alternatively, we could to add a pragma GCC diagnostic pop at the end of
this file, but this change is simpler and doesn't seem to be any worse.