Add PETE vector tests
Brought
  tools/pete-2.1.1/et_vector_test.cpp
into lmi's unit-test suite as
  et_vector_0_test.cpp
with only minimal changes:
 - updated copyrights to reflect all modification years
 - added lmi unit-test boilerplate
 - worked around useless-cast warnings

The '_0' differentiates this file from the original in tools/ ,
and allows for more types of tests to be added later--e.g., for
other expression-template libraries.

'timer.o' is linked because it'll probably be wanted eventually,
though it's not actually used now.