Remove tests involving std::mem_fun
std::mem_fun and the related types were deprecated in C++11 and
removed in C++17, so it doesn't make sense to have them in lmi which
requires C++17 now.

In practice, MSVS and gcc still provide std::mem_fun() even in C++17
mode, but clang does not (as it has every right to do), so removing this
code fixes the test compilation with clang.