Avoid using gcc-specific warnings with clang
Because clang predefines __GNUC__ (being mostly compatible with gcc),
pragmas inside "#if defined __GNUC__" are also used with clang. But
because clang is not perfectly compatible with the recent gcc versions,
it complains about unknown warnings.

Fix this by explicitly checking for "gcc and not clang". A perhaps
better alternative could be to predefine LMI_GCC symbol for the real gcc
and LMI_GCC_OR_CLANG for the code which should be used with both
compilers.