Make an improvement deferred since 20100428T1231Z
This improvement was mentioned in commit ec606b9c7577 of 20100428T1231Z
but deferred because it would have caused many minuscule regressions
at any point before the immediately preceding commit--but it causes no
regression now that DCV calculations are performed in whole cents.

Of course, that's not a guaranteed effect of such quantization, just a
highly probable one.