Make 7702 statutory rates scalar
This revision causes regressions in 'Irc7702NetGSP' only, the largest
|absolute| and |relative| errors being 8.67362e-19 and 1.85737e-16,
respectively. That's 1 ulp, DBL_EPSILON being 2.2204460492503131E-16.

It's plausible that GSP is affected while GLP is not because the
statutory rate is 'X' in these formulas:
   GLP = Max(       X, Y) - Z
   GSP = Max(0.02 + X, Y) - Z
and 'Y' may dominate in the first but not the second. Alternatively,
given that X is 0.04 in all affected testcases, perhaps 0.04 rounds
one way while 0.06 rounds another way when popped from an 80-bit x87
register, and changing X from vector to scalar removes a needless
round trip through the x87.