Fix defect introduced 20050114T1947Z: wrong DCV NAAR discount
The DCV death benefit must be discounted at a rate that takes the
statutory rate into account. The SOA textbook (on page 52 of the first
edition) offers a questionable rationale for treating it as "not an
interest rate guaranteed on issue of the contract" (i.e., not subject
to the rule in §7702(b)(2)(A)), yet it is clearly an interest rate,
and it is clearly guaranteed in the contract.

Although this defect was introduced 20050114T1947Z into CVS, it comes
from an original file predating the lmi epoch.