Create required files in the build directory with autotools
The existing tests rely on existence of these files in the current
directory and they can't (easily) be run from the source directory with
automake, so create the required files in the build directory instead.

A better solution might be to make the directory with these files
configurable, e.g. via an environment variable, but for now this is
enough to allow "make check" to work.