Fix Boost detection and use in configure with clang in C++17 mode
The use of "register" in one of the headers triggers a -Wregister which
is actually a fatal error, so get rid of it by simply removing
"register" using preprocessor.

Also define BOOST_NO_AUTO_PTR to prevent the use of std::auto_ptr<>
which is not available in C++17 version of libc++. Unlike the "register"
hack, this is also needed for compiling lmi code using this header.