Disable clang -Wstring-plus-int warning
This warning is given for all the occurrences of multiline raw strings
starting, according to lmi convention, as

	1 + R"--cut-here--(
	...
	)--cut-here--";

due to the use of "1 +" and there doesn't seem to be any way to avoid it
in this case without disabling it globally.

As this warning doesn't seem that useful, do just disable it.