Restore but rename PETE's operator==() and !=() for std::vector
Reworked commit 2844ba037f4114 of 20210216T1619Z. Instead of removing
all of PETE's equality and inequality operators:
  operator==(vector, vector)     (1)
  operator==(vector, T)          (2a)
  operator==(T, vector)          (2b)
  operator==(vector, expression) (3a)
  operator==(expression, vector) (3b)
simply renamed them. Element-wise [in]equality comparisons are useful,
though probably not so often that they must be operators as such.

Chose names "EqualTo" and "UnequalTo" for unambiguity, not terseness.

They're still referred to as "operators" in the PETE code, e.g.:
  m["binaryBoolOps"].push_back(
    OperatorDescriptor("OpEQ",
but those are just arbitrary text strings.

It might have been nicer to keep them as operators, suppressing them
only to the extent they ambiguate the important standard operators like

  template<typename T, typename Alloc>
  constexpr bool operator==
    (std::vector<T,Alloc> const& lhs
    ,std::vector<T,Alloc> const& rhs
    );

but that turned out not to be worth the trouble:

(A) Mangling (1) above only, thus:
  sed -i $@ -e's/operator==\(.*std::vector.*std::vector\)/XyZzY_EQ\1/'
  sed -i $@ -e's/operator!=\(.*std::vector.*std::vector\)/XyZzY_NE\1/'
was insufficient because (2a) and (2b) still ambiguated the standard
operator. The 'sed' technique was tried because changing PETE to prevent
the operators from being generated is too daunting: PETE has several
layers of abstraction, and by the time the to-be-suppressed code is
generated, the identity of individual operators has been erased, so a
deep restructuring would be required.

(B) SFINAE might be used, e.g.:

  template<typename>
  struct is_std_vector : std::false_type {};

  template<typename T, typename A>
  struct is_std_vector<std::vector<T,A>> : std::true_type {};

  template<typename T0, typename T1, typename=is_std_vector<T1>::value>
  bool operator==(std::vector<T0> t0, T1 t1) {return t0 == t1;}

but intruding an extra template parameter into PETE seemed similarly
daunting for similar reasons.

(C) Later, of course, outside of PETE's code generation, SFINAE-based
operators could be implemented in terms of "EqualTo" and "UnequalTo",
if desired.

* tools/pete-2.1.1/PETE/Tools/PeteOps.in: Edited.
* tools/pete-2.1.1/PETE/OperatorTags.h: Regenerated.
* tools/pete-2.1.1/PETE/Tools/PeteOps.cpp: Regenerated.
* tools/pete-2.1.1/et_vector_operators.hpp: Regenerated.