Efficiently determine whether a vector<double> is all boolean-valued
For x86_64-pc-linux-gnu gcc-10.2, PETE is fastest:
  mete_eq0a 2.024e-07 s mean;          0 us least of 49399 runs
  mete_eq0b 2.014e-07 s mean;          0 us least of 49661 runs
  mete_eq0c 1.745e-07 s mean;          0 us least of 57313 runs
  mete_eq1a 1.408e-07 s mean;          0 us least of 71035 runs
  mete_eq1b 1.423e-07 s mean;          0 us least of 70268 runs
  mete_eq1c 2.118e-07 s mean;          0 us least of 47224 runs
  mete_eq1x 2.038e-07 s mean;          0 us least of 49067 runs
while for i686-w64-mingw32 gcc-8.3, all take about the same time.

The crucial conclusion is that writing code as close as possible to
C++ extended with an APL declaration like the standard asm-declaration:
  APL {"∧/X∨.=0 1"};   // More idiomatic. (APL '=' means C '=='.)
  APL {"∧/(X=1)∨X=0"}; // Doesn't require constructing vector{0, 1}.
by using PETE
  ∧/   (X=1)                      ∨  X=0
  AllOf(EqualTo(true, bit_valued) || EqualTo(false, bit_valued));
incurs no speed penalty in the particular case studied (determining
whether a 100-element vector<double> is all boolean-valued).

It's curious that
  ∧/(X=1)∨(X=0)  ⍝ all are either =1 or =0
is faster than
  ~∨/(X≠1)∧(X≠0) ⍝ none is both   ≠1 and ≠0
and perhaps less surprising that
  ∧/X=~~X        ⍝ each equals its boolean NOT(NOT)
is not the fastest.