Move an assertion upstream, and generalize it
All 'C..._' must be nonzero after the initial duration. Asserted so
in assert_preconditions() rather than in only one constructor. Moved
associated documentation to a more natural place, rewording it, in
particular to explain why a vector that must be uniformly zero after
the initial duration is preferable to a scalar representing only the
initial value. The reason, if not obvious, is that a universal formula
  max(A, Bgen, Cgen) - Dgen
is clearer than one with conditional alternatives
  max(A, Bgen, Cgen) - Dgen  (first duration only)
  max(A, Bgen      ) - Dgen  (all other durations)