Assert more preconditions
If all 'use_..._' are false in any year, then the formula
   max
     ((max(A, Bgen, Cgen) - Dgen) * use_gen
     ,(max(A, Bsep, Csep) - Dsep) * use_sep
     ,(max(A, Bflr, Cflr) - Dflr) * use_flr
     ,(max(A, Bvlr, Cvlr) - Dvlr) * use_vlr
     )
would return zero, but the law requires that it return at least 'A'.

Rewriting the formula to guarantee that directly:
   max
     (A
     ,max
       ((max(A, Bgen, Cgen) - Dgen) * use_gen
       ,(max(A, Bsep, Csep) - Dsep) * use_sep
       ,(max(A, Bflr, Cflr) - Dflr) * use_flr
       ,(max(A, Bvlr, Cvlr) - Dvlr) * use_vlr
       )
     )
would make it more complicated and slower. Presumably at least one of
{use_gen_, use_sep_} must always be true in the problem domain anyway,
so that the assertion would never fire.