Prefer PETE's operator<<=() to assign()
Replaced assign() and the awful
  std::vector<double> y(x.size());
  x += y;
anti-idiom with operator<<=(), which, despite its ugliness, is less bad.

This change does make mete_reserve() slower, by about ten percent for
x86_64-pc-linux-gnu and a hundred percent for i686-w64-mingw32, but
that's okay because the penalty is due to heap allocation. Assigning
PETE expressions to a vector that already has the right size is faster
than replacing the vector, but that's a PETE matter for separate
investigation. Here, the goal is to write complex code clearly. The
reserve calculation is apparently so fast that allocating a new vector
on the heap is not vanishingly cheap in comparison. The slowness of
small heap allocations for i686 on 'wine' is the tail that mustn't be
allowed to wag the dog.