Do away with an undesirable smart pointer [290]
Sometimes using unique_ptr simplifies the physical design by making a
header smaller after preprocessing. Here, the tradeoff was not clearly
favorable: including <memory>, which is not insignificant, vs. the
definitions of two tiny classes and a couple of minor headers. Using
unique_ptr had also required writing a defaulted dtor out of line, and
accessing members with '->' instead of the simpler '.'.