Add an inchoate test of power-of-ten scaling
Compare numerical results of
  scaled = (1e8 * unscaled) / 1e8
  scaled = (1e8 * unscaled) * (1 / 1e8)
using e as the value of 'unscaled', caching the reciprocal both
as double and as long double.

The result is a lossless round trip for i686, and for x86_64 with a
long double reciprocal but not with a double reciprocal. Changing the
'unscaled' value can produce different outcomes; e.g., with 1.0, 4/3,
and 10/9, all results were lossless round trips.