Remove a source of inaccuracy [285]
Formerly, scale_fwd_ was something like 1000 or 0.001, and scale_back_
was its reciprocal, e.g., 1/1000 or 1/.001; but calculating 1/.001, the
reciprocal of the reciprocal of 1000, invites error. Now, they're e.g
1000 and 1/1000, or 1/1000 and 1000. This may not matter as long as the
factors use extended precision, but it's a needless impediment to the
possible future use of double precision.

Expunged an unnecessary cover function for nonstd::power(). It was not
used in enough places to warrant a separate function; worse, it masked
unintended outcomes like calculating 1/.001 above.

Abandoned the idea of favoring const data members. It is more commonly
advised that they be avoided. The idea that they would permit better
optimization seems implausible, and is hardly relevant because round_to
objects are used far more frequently than they're constructed.