Add and use a modal Nx UL commutation function
This documentation in 'commutation_functions.hpp' always anticipated
providing a modal Nx:

/// All commutation functions are calculated on the mode specified
/// by mode_. Annual D and N are always also calculated because
/// premiums are often paid annually. Use monthly D and N for
/// monthly deductions in the numerator of an actuarial function,
/// but use their annual analogs in the denominator when premiums
/// are assumed to be paid annually.

Nx had never previously been wanted, because, given a vector V and a
scalar S = V[x], N[x] times S is just shorthand for the dot product
D·V in the common but not universal case that all V[k] are equal.
Now Nx is handy for a textbook example that assumes the common case.