Wait for Xvfb to become available before running the GUI test
In the CI environment, the test can be launched before the Xvfb server
becomes fully available, resulting in X connection errors and failure to
create any windows.

Work around this by waiting until we can connect to the server before
running the test. Waiting just once for 3 seconds seems to be enough to
avoid the problem, but wait for up to 30 seconds just to make sure that
if we fail, it's not due to the slow Xvfb startup.