Another work around for spurious GUI test failures in CI builds
Try adding an extra delay on top of the wait loop added in 87a7956f3
(Wait for Xvfb to become available before running the GUI test,
2021-04-16) as we still get Wine nodrv_CreateWindow errors when trying
to use Xvfb even when connecting to it already works.

This is ugly, but there doesn't seem to be any way of checking if using
Xvfb already works other than building and launching a test program just
for this, which looks like an overkill -- but will still probably be the
next step, if even this delay proves to be insufficient.