Favor designated initializers
The next commit will rearrange the order of gpt_scalar_parms members.
What would happen if this unit test were not correspondingly changed?
Without the present commit, it would compile successfully; it would
fail at run time:
  Assertion 'a_SpecAmt <= a_BftAmt' failed.
but only because lmi has so many assertions, one of which happens to
fire (although it's in code that will soon be replaced). With the
present commit, it would fail at compile time:
  designator order for field [...] does not match declaration order
as is much to be preferred.