Add database entities for a minimum single premium
Single-premium sales almost always involve a chosen dollar amount of
premium, and a specified amount calculated as a function of that chosen
premium: the corresponding CVAT corridor, or perhaps some percentage
such as 95% of corridor.

For example, if the premium is $10000 and the corridor factor is 2.02,
then the specified amount would be entered as "corridor" and lmi would
realize that strategy as $20200.

But what happens if the minimum specified amount is, say, $50000? In the
example above, lmi increases the specified amount to $50000...and leaves
the customer's premium alone, because it would be astonishing to demand
$20200 when only $10000 was offered. However, this means that the single
premium is inadequate for the (forced) minimum specified amount, so the
illustration lapses early. It would be reasonable to conclude that the
minimum policy size should be stated in terms of premium rather than
specified amount, and amend the policy specifications accordingly.

If policy specifications are deemed to be immutable, the only reasonable
thing lmi can do is refuse to produce an illustration. Therefore, added
two database entities to support such a refusal.

Some multipliers default to unity because zero would be perilous: e.g.,
setting DB_GuarCoiMultiplier to anything less than one is probably a
mistake; to zero, a catastrophe. However, for DB_MinSinglePremiumMult,
the default is zero because there is no higher a priori minimum.