Add an equality operator
It is anticipated that all but one line of this change will be expunged
soon enough.

Unfortunately the shorthand
  default: ==, !=, <, >, <=, >=;
proposed in
  http://www.open-std.org/Jtc1/sc22/wg21/docs/papers/2014/n4126.htm
has not been embraced by C++20.

SD-6 says: "in most cases it is expected that the presence of a feature
can be determined by the presence of any non-zero macro value".