Revert "Work around a boost issue"
This reverts commit 9a631cddf896eeda5c4fc2d8f2817cbcda612918 as the
unexpected exception is not thrown when using std::filesystem.

Notice that the test may still fail in the situation described in the
comment being removed by this commit, i.e. if a path exists but is
inaccessible, because chdir() fails in this scenario. However
is_directory() still returns true, without throwing any exceptions, even
when using an unformatted disk in a drive.