Temporarily move a wine workaround aside
This commit temporarily reverts a part of commit 26dd8aa5c0b, only for
the purpose of merging a set of std::filesystem changes. It doesn't
entirely revert the changes to this single file; it leaves the new
'#include' directive, because there's no need to back that out and
later put it back in.