Record speed measurements [terque]
These measurements use gcc-10 for all architectures, and wine-4.0.3
(wine four, not five). Compared to commit 641831ad4:
  - i686-w64-mingw32 is thirty percent faster for the no-solves
      scenario, and forty to fifty percent faster for other scenarios
  - x86_64-pc-linux-gnu is about the same--maybe one percent faster
  - x86_64-w64-mingw32 is about one percent slower
If the effect of std::filesystem isn't much, then gcc-10 would seem
to have made an enormous different for i686 with wine, and to have
eliminated most of the x86_64-w64-mingw32 vs. i686-w64-mingw32
advantage; it would be interesting to learn whether that happens with
native msw.