Say "basename" rather than "leaf"
The POSIX dirname/basename nomenclature is universally understood.
Boost used "leaf", whose meaning was unmistakable (although a filepath
is hardly a tree). Now std::filesystem uses "filename", which is too
easily misunderstood: in the grammar, a full path can be
  filename / filename
but the filename() function is specified to return '*--end()', so
  filename(filename0 / filename1)
returns filename1, which is "the" filename, though filename0 is also
a filename.