Improve documentation
This change in [squashed] commit e0362c7971676:

-/// fs::change_extension() uses the present argument order:
-///   function(original, new_part)
+/// fs::path::replace_extension() uses the present argument order:
+///   original.replace_extension(new_part)

lost the meaning of the original--i.e., lmi function modify_directory()
takes two arguments in a particular order, and change_extension() in
the boost original took two corresponding arguments in the same order.
Because the standard replace_extension() takes only one argument, there
is no parallel to be drawn.