Rework a 'wine' workaround again
Cf. commits
  26dd8aa5c0b40
  6d8e4f143e069
  fbb5aacf5eaf4

Felicitously, perhaps, some versions of 'wine' happen to choke on a
'....' path, and wind up producing an outcome that is reasonable.
This exotic edge case is not an lmi error and no sane end user would
encounter it. Using the insights recorded here, the whole matter can
be revisited someday if desired.