Revert "Revert "Trap exceptions from filesystem library""
This reverts commit dab074e388422f328933752ae6965560e2c9a62e.

Apparently that commit was intended as a simplification only: it removed
complicated code that couldn't actually have any effect. Normally, that
would be accepted without question. Furthermore, the code it removed
made configurable_settings fragile, as discussed in this thread:
  https://lists.nongnu.org/archive/html/lmi/2020-10/msg00037.html
and, normally, that would cause it to be embraced with enthusiasm.

However, given '/opt/lmi/data/configurable_settings.xml' contents:
  <default_input_filename>Z:/etc/opt/lmi/default.ill</default_input_...
  <print_directory>Z:/opt/lmi/print</print_directory>
lmi-gtk's GUI now shows
  Input defaults: /opt/lmi/bin/Z:/etc/opt/lmi/default.ill
  Printout directory: /opt/lmi/print
where one textcontrol is as intended but the other is not. It is not
clear whether the reverted commit has anything to do with that, but
it seems wise to back it out temporarily pending investigation of this
anomaly. The present commit can be reverted (enthusiastically) later.