Clarify unit-test documentation
It does seem unfortunate that boost's free function
  path change_extension(path const& p, std::string const& new_ext);
was standardized only as non-const member
  path& replace_extension(path const& new_ext = path());
because almost half of lmi's replace_extension() calls
  $git grep replace_extension
must take extra care to preserve the argument, e.g.:
  fs::path{cell_filepath}.replace_extension(".test")

It is tempting to add such a free function, not only for convenience
but also for safety. For now, hypothesized that to have been done,
and rewrote documentation to be less confusing.