Amend 7702 specifications: (f)(1)(A) amount can be negative
The phrase "premiums paid" in the original clearly meant tax basis (in
contradistinction to income in the contract), and not 'premiums paid' as
defined by the statute. Replaced the misleading phrase.

For example, if cum_prems_paid was $30, and the guideline_limit becomes
-$100 due to an adjustment event, then

    forceout_amount = cum_prems_paid - guideline_limit
    cum_prems_paid = cum_prems_paid - forceout_amount

    forceout_amount = 30 - (-100) = 130
    cum_prems_paid = 30 - 130 = -100

so a negative guideline limit causes (f)(1)(A) 'premiums paid' to become
negative.