Make an operator 'consteval'
As shown in the unit test, any misuse is now a compile-time error, even
outside a constexpr context.

Incidentally simplified the error message, which can appear only at
compile time, e.g.:

in ‘constexpr’ expansion of ‘operator""_cents(9007199254740993)’
error: expression ‘<throw-expression>’ is not a constant expression
   79 |         : throw std::domain_error("outside currency domain")
      |           ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

so extra code to print the value a second time is unwarranted.