Work around std::allocator::size_type problem in clang 11 libc++
This (and only this) version of libc++ has apparently mistakenly removed
std::allocator::size_type definition, which is still present in C++20,
and requires predefining a special symbol in order to do it -- which we
must do to compile boost/regex.hpp as it (indirectly) uses this type.

Add a configure test checking if we really need to do it and do define
it if it is required.