Create /tmp/.X11-unix directory before running Xvfb
This avoids the following warning

_XSERVTransmkdir: ERROR: euid != 0,directory /tmp/.X11-unix will not be
created.

when running it, which seems to be harmless and doesn't prevent the
tests from passing, but it still seems better not to have it.

Notice that this directory must have 1777 permissions (and not just 777
or 755) to avoid another warning from X server about this.