Fix test_path_validation() for compilers without char8_t support
This makes the test pass with clang-11 and, more generally, makes it
more robust by not relying on having the part of the string following
the first NUL character in the exception error message, as this message
is constructed using std::runtime_error::what() returning "const char*"
and so is always terminated by the first NUL in it.

Note that previously the test still passed with the compilers that do
support char8_t because of the compensating buglet in fs::path, whose
u8string_as_string() also truncated the string at the first NUL, and did
it before constructing std::runtime_error(), thus ensuring that the
string passed to it didn't contain any NULs any longer and so, while the
part of the path after the first NUL was still lost (even if this was
difficult to see with a path containing only NULs, as was previously
used in the test, which is the reason for adding non-NUL characters to
it too now), the rest of the string could be recovered later.